requirements.txt
pytz
requests
python-dotenv
ipython

# Django
django
django-timezone-field 
django-cors-headers
django-debug-toolbar
whitenoise

# Django 3.0 async requirements
gunicorn
uvicorn
#uvloop
httptools

# Databases
psycopg2-binary
dj-database-url

# Distributed tasks
celery
redis
django-celery-beat

# Telegram
python-telegram-bot==13.13

# monitoring
# sentry-sdk
